Reference

BondGraphs.Component

BondGraphs.Component

BondGraphs.EqualEffort

BondGraphs.EqualFlow

BondGraphs.SourceSensor

BondGraphs.connect!

BondGraphs.Component — Type

Component(type, name=type)
Component(type, name=type; library=BondGraphs.DEFAULT_LIBRARY, <keyword arguments>)

Construct a Component of a defined (bondgraph) type ∈ {R, C, I, Se, Sf, TF, Ce, Re, SCe}. Properties and equations of available componets are defined in BondGraphs.DEFAULT_LIBRARY.

BondGraphs.Component — Type

Component{N} <: AbstractNode

N-port bond graph component.

BondGraphs.EqualEffort — Type

EqualEffort <: Junction

Efforts are all equal, flows sum to zero (aka 0-junction).

BondGraphs.EqualFlow — Type

EqualFlow <: Junction

Flows are all equal, efforts sum to zero (aka 1-junction).

BondGraphs.SourceSensor — Type

SourceSensor <: AbstractNode

Special component type that acts as a source of both effort and flow

BondGraphs.connect! — Method

connect!(bg, srcnode, dstnode)

Connect two components together in the same bond graph.